+++ src/Xbehave.Sdk.Net35/Xbehave.Sdk.Net35.csproj
@@ -97,9 +97,6 @@
     <Compile Include="..\Xbehave.Sdk.Net40\ReplayCommand.cs">
       <Link>ReplayCommand.cs</Link>
     </Compile>
-    <Compile Include="..\Xbehave.Sdk.Net40\Step.AsyncTestSyncContext.cs">
-      <Link>Step.AsyncTestSyncContext.cs</Link>
-    </Compile>
     <Compile Include="..\Xbehave.Sdk.Net40\Step.cs">
       <Link>Step.cs</Link>
     </Compile>
