+++ src/Xbehave.Sdk.Net45/AsyncStep.cs
@@ -5,37 +5,50 @@
 namespace Xbehave.Sdk
 {
     using System;
-    using System.Threading;
     using System.Threading.Tasks;
+    using Xunit.Sdk;
 
     public class AsyncStep : Step
     {
         private readonly Func<Task> body;
-        private Task task;
 
         public AsyncStep(string name, Func<Task> body, object stepType)
             : base(name, stepType)
         {
+            Guard.AgainstNullArgument("body", body);
+
             this.body = body;
         }
-        
-        protected override void ExecuteBody()
-        {
-            this.task = this.body.Invoke();
-        }
 
-        protected override Exception WaitForCompletion()
+        public override void Execute()
         {
             try
             {
-                this.task.Wait();
+                if (this.MillisecondsTimeout > 0)
+                {
+                    var task = this.body();
+                    var timeout = Task.Delay(this.MillisecondsTimeout);
+                    if (Task.WhenAny(task, timeout).Result == timeout)
+                    {
+                        throw new Xunit.Sdk.TimeoutException(this.MillisecondsTimeout);
+                    }
+                }
+                else
+                {
+                    this.body().Wait();
+                }
             }
-            catch (AggregateException ae)
+            catch (AggregateException ex)
             {
-                return ae.InnerException;
+                ExceptionUtility.RethrowWithNoStackTraceLoss(ex.InnerException);
             }
-
-            return null;
+            finally
+            {
+                foreach (var teardown in this.Teardowns)
+                {
+                    CurrentScenario.AddTeardown(teardown);
+                }
+            }
         }
     }
 }
