+++ src/Xbehave.Sdk.Net45/Xbehave.Sdk.Net45.csproj
@@ -98,9 +98,6 @@
     <Compile Include="..\Xbehave.Sdk.Net40\ReplayCommand.cs">
       <Link>ReplayCommand.cs</Link>
     </Compile>
-    <Compile Include="..\Xbehave.Sdk.Net40\Step.AsyncTestSyncContext.cs">
-      <Link>Step.AsyncTestSyncContext.cs</Link>
-    </Compile>
     <Compile Include="..\Xbehave.Sdk.Net40\Step.cs">
       <Link>Step.cs</Link>
     </Compile>
